+++ src/main/java/com/wondertek/cpm/web/rest/UserCostResource.java
@@ -1,12 +1,13 @@
 package com.wondertek.cpm.web.rest;
 
-import com.codahale.metrics.annotation.Timed;
-import com.wondertek.cpm.domain.UserCost;
-import com.wondertek.cpm.service.UserCostService;
-import com.wondertek.cpm.web.rest.util.HeaderUtil;
-import com.wondertek.cpm.web.rest.util.PaginationUtil;
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.time.ZonedDateTime;
+import java.util.List;
+import java.util.Optional;
 
-import io.swagger.annotations.ApiParam;
+import javax.inject.Inject;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Page;
@@ -14,17 +15,28 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.security.access.annotation.Secured;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
-import javax.inject.Inject;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
-import java.util.stream.StreamSupport;
+import com.codahale.metrics.annotation.Timed;
+import com.wondertek.cpm.CpmConstants;
+import com.wondertek.cpm.config.StringUtil;
+import com.wondertek.cpm.domain.UserCost;
+import com.wondertek.cpm.security.AuthoritiesConstants;
+import com.wondertek.cpm.security.SecurityUtils;
+import com.wondertek.cpm.service.UserCostService;
+import com.wondertek.cpm.web.rest.util.HeaderUtil;
+import com.wondertek.cpm.web.rest.util.PaginationUtil;
 
-import static org.elasticsearch.index.query.QueryBuilders.*;
+import io.swagger.annotations.ApiParam;
 
 /**
  * REST controller for managing UserCost.
@@ -47,6 +59,7 @@ public class UserCostResource {
      */
     @PostMapping("/user-costs")
     @Timed
+    @Secured(AuthoritiesConstants.ROLE_INFO_USERCOST)
     public ResponseEntity<UserCost> createUserCost(@RequestBody UserCost userCost) throws URISyntaxException {
         log.debug("REST request to save UserCost : {}", userCost);
         if (userCost.getId() != null) {
@@ -69,15 +82,60 @@ public class UserCostResource {
      */
     @PutMapping("/user-costs")
     @Timed
+    @Secured(AuthoritiesConstants.ROLE_INFO_USERCOST)
     public ResponseEntity<UserCost> updateUserCost(@RequestBody UserCost userCost) throws URISyntaxException {
         log.debug("REST request to update UserCost : {}", userCost);
-        if (userCost.getId() == null) {
-            return createUserCost(userCost);
+        Boolean isNew = null;
+        if(userCost == null || userCost.getUserId() == null || userCost.getCostMonth() == null || 
+        		userCost.getUserName() == null || userCost.getStatus() == null){
+    		return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.userCost.save.requriedError", "")).body(null);
+    	}
+        //获取当前用户
+        String updator = SecurityUtils.getCurrentUserLogin();
+        ZonedDateTime updateTime = ZonedDateTime.now();
+        isNew = userCost.getId() == null;
+        UserCost findUserCost = null;
+        if(isNew){//新增
+        	userCost.setId(null);
+        	
+        	findUserCost = userCostService.findByUserIdAndCostMonth(userCost.getUserId(),userCost.getCostMonth());
+        	if(findUserCost != null){
+        		return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.userCost.save.existError", "")).body(null);
+        	}
+        	findUserCost = new UserCost();
+        	findUserCost.setUserId(userCost.getUserId());
+        	findUserCost.setUserName(userCost.getUserName());
+        	findUserCost.setCostMonth(userCost.getCostMonth());
+        	findUserCost.setCreateTime(updateTime);
+        	findUserCost.setCreator(updator);
+        }else{//编辑
+        	findUserCost = userCostService.findByUserIdAndCostMonth(userCost.getUserId(),userCost.getCostMonth());
+        	if(findUserCost == null){
+        		return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.userCost.save.noExistError", "")).body(null);
+        	}else if(userCost.getUserName() == null || userCost.getStatus() == null){
+        		return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.userCost.save.requriedError", "")).body(null);
+        	}else if(findUserCost.getStatus() == CpmConstants.STATUS_DELETED){
+        		return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.userCost.save.statusError", "")).body(null);
+        	}else if(!userCost.getUserName().equals(findUserCost.getUserName())){
+        		return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.userCost.save.userNameError", "")).body(null);
+        	}
         }
-        UserCost result = userCostService.save(userCost);
-        return ResponseEntity.ok()
-            .headers(HeaderUtil.createEntityUpdateAlert("userCost", userCost.getId().toString()))
-            .body(result);
+        findUserCost.setStatus(userCost.getStatus());
+        findUserCost.setExternalCost(userCost.getExternalCost());
+    	findUserCost.setInternalCost(userCost.getInternalCost());
+    	findUserCost.setUpdateTime(updateTime);
+    	findUserCost.setUpdator(updator);
+    	
+        UserCost result = userCostService.save(findUserCost);
+        if(isNew){
+        	return ResponseEntity.created(new URI("/api/user-costs/" + result.getId()))
+                    .headers(HeaderUtil.createEntityCreationAlert("userCost", result.getId().toString()))
+                    .body(result);
+        }else{
+        	return ResponseEntity.ok()
+                    .headers(HeaderUtil.createEntityUpdateAlert("userCost", userCost.getId().toString()))
+                    .body(result);
+        }
     }
 
     /**
@@ -89,15 +147,30 @@ public class UserCostResource {
      */
     @GetMapping("/user-costs")
     @Timed
+    @Secured(AuthoritiesConstants.ROLE_INFO_USERCOST)
     public ResponseEntity<List<UserCost>> getAllUserCosts(
-    		@RequestParam(value = "userId",required=false) Long userId,
-    		@RequestParam(value = "userName",required=false) Long userName,
-    		@RequestParam(value = "costMonth",required=false) Long costMonth,
-    		@RequestParam(value = "status",required=false) Long status,
+    		@RequestParam(value = "userId",required=false) String userId,
+    		@RequestParam(value = "userName",required=false) String userName,
+    		@RequestParam(value = "costMonth",required=false) String costMonth,
+    		@RequestParam(value = "status",required=false) String status,
     		@ApiParam Pageable pageable)
         throws URISyntaxException {
         log.debug("REST request to get a page of UserCosts");
-        Page<UserCost> page = userCostService.findAll(pageable);
+        UserCost userCost = new UserCost();
+        if(!StringUtil.isNullStr(userId)){
+        	userCost.setUserId(StringUtil.nullToLong(userId));
+        }
+        if(!StringUtil.isNullStr(userName)){
+        	userCost.setUserName(userName);
+        }
+        if(!StringUtil.isNullStr(costMonth)){
+        	userCost.setCostMonth(StringUtil.nullToLong(costMonth));
+        }
+        if(!StringUtil.isNullStr(status)){
+        	userCost.setStatus(StringUtil.nullToInteger(status));
+        }
+        Page<UserCost> page = userCostService.getUserCostPage(userCost,pageable);
+//        Page<UserCost> page = userCostService.findAll(pageable);
         HttpHeaders headers = PaginationUtil.generatePaginationHttpHeaders(page, "/api/user-costs");
         return new ResponseEntity<>(page.getContent(), headers, HttpStatus.OK);
     }
@@ -110,6 +183,7 @@ public class UserCostResource {
      */
     @GetMapping("/user-costs/{id}")
     @Timed
+    @Secured(AuthoritiesConstants.ROLE_INFO_USERCOST)
     public ResponseEntity<UserCost> getUserCost(@PathVariable Long id) {
         log.debug("REST request to get UserCost : {}", id);
         UserCost userCost = userCostService.findOne(id);
@@ -128,8 +202,13 @@ public class UserCostResource {
      */
     @DeleteMapping("/user-costs/{id}")
     @Timed
+    @Secured(AuthoritiesConstants.ROLE_INFO_USERCOST)
     public ResponseEntity<Void> deleteUserCost(@PathVariable Long id) {
         log.debug("REST request to delete UserCost : {}", id);
+        UserCost userCost = userCostService.findOne(id);
+        if(userCost.getStatus() == CpmConstants.STATUS_DELETED){
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.userCost.delete.statusError", "")).body(null);
+        }
         userCostService.delete(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert("userCost", id.toString())).build();
     }
@@ -145,6 +224,7 @@ public class UserCostResource {
      */
     @GetMapping("/_search/user-costs")
     @Timed
+    @Secured(AuthoritiesConstants.ROLE_INFO_USERCOST)
     public ResponseEntity<List<UserCost>> searchUserCosts(@RequestParam String query, @ApiParam Pageable pageable)
         throws URISyntaxException {
         log.debug("REST request to search for a page of UserCosts for query {}", query);
